structure of the project
