+++ .gitignore
@@ -3,3 +3,4 @@
 *.pem
 *.env
 node_modules/
+.DS_Store
