+++ package.json
@@ -1,10 +1,16 @@
 {
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "node index.js",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
     "dotenv": "^16.4.0",
     "express": "^4.18.2"
+  },
+  "devDependencies": {
+    "vite": "^6.3.5"
   }
 }
