--- public/home.html
@@ -1,148 +1,0 @@
-<!DOCTYPE <html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./css/reset.css">
-    <link rel="stylesheet" href="./css/home.css">
-    <title>ABBY</title>
-</head>
-
-<body>
-    <header>
-        <img src="./assets/img/logo.png" alt="Abby logo" width="150">
-        <p>EN</p>
-        <p>profile icon</p>
-    </header>
-    <main>
-        <section class="intro">
-            <h1>Hey! I’m <span class="abby_name_color">Abby</span></h1>
-            <p class="intro_text">I made this <strong>game</strong> for anyone figuring out <strong>living on their
-                    own</strong>, like me!</p>
-            <div class="abby">
-                <svg xmlns="http://www.w3.org/2000/svg" width="104" height="81" viewBox="0 0 104 81" fill="none">
-                    <path d="M104 81L74.4065 0H0L104 81Z" fill="white" />
-                </svg>
-                <img class="abby_hi" src="./assets/img/abby-hi.png" alt="Abby character">
-            </div>
-            <div class="cta">
-                <p>Are you already at Abby’s?</p>
-                <button>Let's Play</button>
-            </div>
-            <div class="intro_extra">
-                <p class="intro_extra_text">Abby is navigating solo life at the new museum. Come explore and
-                    <strong>grow
-                        together.</strong></p>
-            </div>
-        </section>
-        <section class="about">
-            <h2 class="hidden">About</h2>
-            <div class="box about_who">
-                <img src="./assets/svg/who.svg" alt="icon">
-                <h3>Who?</h3>
-                <p>Everyone is welcome to grow with Abby!</p>
-            </div>
-            <div class="box about_where">
-                <img src="./assets/svg/where.svg" alt="icon">
-                <h3>Where?</h3>
-                <p>We’re at Begijnhofpark Museum’s open spaces.</p>
-                <a href="#">Open in Google Maps &#128279;</a>
-            </div>
-            <p>It's a gentle way to <strong>connect</strong> and <strong>grow together</strong> with all my
-                museum-friends.
-            </p>
-            <div class="box about_how">
-                <img src="./assets/svg/how.svg" alt="icon">
-                <h3>How?</h3>
-                <p>Connect with fellow visitors, answer questions and share your experience</p>
-            </div>
-            <div class="box about_when">
-                <img src="./assets/svg/when.svg" alt="icon">
-                <h3>When?</h3>
-                <p>We’re open Tuesday to Sunday <strong>10AM</strong> to <strong>10PM</strong></p>
-            </div>
-        </section>
-        <section class="events">
-            <h2>Upcoming events</h2>
-            <div>
-                <div class="event">
-                    <h3>Game night</h3>
-                    <p>Cozy talks and new connections.</p>
-                    <div class="link">
-                        <img src="./assets/svg/arrow.svg" alt="arrow">
-                        <p class="link_text">Read more</p>
-                    </div>
-                </div>
-                <hr>
-                <div class="event">
-                    <h3>Dinner for one</h3>
-                    <p>Simple recipes, shared kitchens.</p>
-                    <div class="link">
-                        <img src="./assets/svg/arrow.svg" alt="arrow">
-                        <p class="link_text">Read more</p>
-                    </div>
-                </div>
-                <hr>
-                <div class="event">
-                    <h3>Money? Let's Try</h3>
-                    <p>What even is a tax return?</p>
-                    <div class="link">
-                        <img src="./assets/svg/arrow.svg" alt="arrow">
-                        <p class="link_text">Read more</p>
-                    </div>
-                </div>
-                <hr>
-            </div>
-            <div>
-                <button>+</button>
-                <button>Explore more</button>
-                <button>Co-create</button>
-            </div>
-        </section>
-        <section class="spaces">
-            <h2>Open spaces</h2>
-            <p class="spaces_text">Welcome to my little world. Here’s my house... Well, our house, really.</p>
-            <div class="space">
-                <div class="space_nav">
-                    <div>
-                        <h3 class="space_name">The Living</h3>
-                        <img class="space_icon" src="./assets/svg/living.svg" alt="icon">
-                    </div>
-                    <div class="space_btns">
-                        <img class="prev_btn" src="./assets/svg/prev.svg" alt="arrow next">
-                        <img class="next_btn" src="./assets/svg/next.svg" alt="arow previous">
-                    </div>
-                </div>
-                <div class="space_about">
-                    <video src="./assets/video/atelier.mp4" autoplay loop muted playsinline class="space_media">
-                    </video>
-                    <!-- <img src="./assets/video/atelier.mp4" class="space_media"> -->
-                    <p class="space_description">We have the living room, where I toss down my groceries, plop myself
-                        anywhere, and honestly just…
-                        exist.</p>
-                    <div class="link">
-                        <img src="./assets/svg/arrow.svg" alt="arrow">
-                        <p class="link_text">Read more</p>
-                    </div>
-                </div>
-            </div>
-            <button>Explore more</button>
-        </section>
-        <section class="people">
-            <h2>People Growing</h2>
-            <div>
-                <h3>Lisa</h3>
-                <p>profile pic</p>
-                <p>Laundry Warrior</p>
-                <p>Dish Duty Ninja</p>
-                <p>Pnaic Plumber</p>
-                <button>See more</button>
-                <button>+</button>
-            </div>
-        </section>
-    </main>
-    <script type="module" src="./scripts/home.js"></script>
-</body>
-
-</html>
