+++ public/intro.html
@@ -1,0 +1,112 @@
+<!DOCTYPE <html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" href="./css/reset.css">
+    <link rel="stylesheet" href="./css/game.css">
+    <title>ABBY</title>
+</head>
+
+<body>
+    <header>
+        <img src="./assets/img/logo.png" alt="Abby logo" width="150">
+        <p>EN</p>
+        <p>profile icon</p>
+    </header>
+    <main>
+        <section class="introduction">
+            <h1 class="hidden">Introduction</h1>
+            <div class="header">
+                <button>back</button>
+                <div>
+                    <h2>Welcome!</h2>
+                    <p>1/4</p>
+                </div>
+            </div>
+            <div class="steps_bg">
+                <div class="hidden step step_one">
+                    <h2>How this works</h2>
+                    <div class="intro">
+                        <h3>Why</h3>
+                        <p>You’ve just started living on your own and so has Abby. Let’s grow side by side Abby and her
+                            friends</p>
+                        <h3>How</h3>
+                        <p>You’ll find <strong>10 activities</strong> on the <strong>map</strong>. Complete them all and
+                            explore freely.</p>
+                        <h3>Extra</h3>
+                        <p>Grab a pen or pencil at the entrance (you’ll need it for some activities!)</p>
+                    </div>
+                    <div class="info">
+                        <p>You can always get the info by tapping the <img src="./assets/svg/info.svg"
+                                alt="information icon"> icon on the right side of the screen!
+                        </p>
+                        <button>Got it!</button>
+                    </div>
+                </div>
+                <div class="step step_two">
+                    <div>
+                        <div class="speech_bg">
+                            <p class="speech">Hi, I'm Abby. Did you know if you mix up my name, it spells ‘baby’? I’ve
+                                always been about
+                                growing.</p>
+                        </div>
+                        <img class="abby_hi" src="./assets/img/abby-hi.png" alt="Abby character">
+                    </div>
+                    <p>What should I call my new friend?</p>
+                    <img src="./assets/svg/question.svg" alt="more info icon">
+                    <input type="text" placeholder="Max the Laundry Conqueror" />
+                </div>
+                <div class="hidden step step_three">
+                    <p>I just started living on my own and I love the freedom like checking art at my own space.
+                        What
+                        aboutyou?</p>
+                    <img class="abby_hi" src="./assets/img/abby-hi.png" alt="Abby character">
+                    <p>What’s something that really makes you smile when you think about living alone?</p>
+                    <button>?</button>
+                    <textarea placeholder="Dancing in my pajamas at midnight"></textarea>
+                </div>
+                <div class=" hidden step step_four">
+                    <p>Although, I dream… and oops, I forget the time! But that’s part of the adventure.</p>
+                    <p>And what’s been a little tricky for you about solo living?</p>
+                    <button>?</button>
+                    <textarea placeholder="Dancing in my pajamas at midnight"></textarea>
+                </div>
+                <button class="next_btn">Next</button>
+            </div>
+        </section>
+        <section class="hidden map">
+            <h2 class="hidden">Map</h2>
+            <button data-mark="l1">living 1</button>
+            <button data-mark="l1">living 2</button>
+            <button data-mark="a1">atelier 1</button>
+            <button data-mark="a1">atelier 2</button>
+            <button data-mark="s1">salon 1</button>
+            <button data-mark="s1">salon 2</button>
+            <button data-mark="c1">cafe 1</button>
+            <button data-mark="c2">cafe 2</button>
+            <button data-mark="g1">garden 1</button>
+            <button data-mark="g2">garden 2</button>
+        </section>
+        <section class="hidden prequestion">
+            <h2 class="hidden">Prequestion</h2>
+            <div class="abby_talk"></div>
+            <ul>
+                <li>Please wait here - another player/s will meet you on the mark.</li>
+                <li>When they arrive, say hello to them.</li>
+                <li>After you’ve said hello, click the button that says: “I found them.”</li>
+            </ul>
+            <button class="foundBtn">I found them!</button>
+        </section>
+        <section class=" hidden extra_info">
+            <h2 class="hidden">One phone</h2>
+        </section>
+        <section class=" hidden question">
+            <h2 class="hidden">Question</h2>
+        </section>
+    </main>
+    <script type="module" src="./scripts/game.js"></script>
+</body>
+
+</html>
