+++ vite.config.js
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+    root: 'public', 
+    build: {
+        outDir: 'dist',
+    },
+});
